+++ test/HspecMain.hs
@@ -12,6 +12,8 @@ import qualified Spec.UnicodeHandling
 import qualified Spec.DefinitionMetadata
 import qualified Spec.DerivedOutput
 import qualified Spec.GoldenWire
+import qualified Spec.HandlerExceptions
+import qualified Spec.HttpTransport
 import qualified Spec.ModernEra
 import qualified Spec.Progress
 import qualified Spec.ProtocolVersionNegotiation
@@ -32,6 +34,8 @@ main = hspec $ do
     Spec.DefinitionMetadata.spec
     Spec.DerivedOutput.spec
     Spec.GoldenWire.spec
+    Spec.HandlerExceptions.spec
+    Spec.HttpTransport.spec
     Spec.ModernEra.spec
     Spec.Progress.spec
     Spec.ProtocolVersionNegotiation.spec